organize folder structure for ca (common, continuous, neural) and implement NCA.forward
